+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project honours [Semantic Versioning](http://semver.org/).
 ------------------------------------------------------------------------
 * **Added:** A new placeholder style for displaying form-feeds as `\f`
 * **Fixed:** Borders missing unless `editor.showInvisibles` is enabled
+* **Changed:** Dividers surrounded by text now visible and shown as `\f`